+++ apps/soongan-api/src/main/kotlin/com/soongan/soonganbackend/soonganapi/interfaces/member/dto/response/MemberInfoResponseDto.kt
@@ -1,6 +1,7 @@
 package com.soongan.soonganbackend.soonganapi.interfaces.member.dto.response
 
 import com.fasterxml.jackson.annotation.JsonInclude
+import com.soongan.soonganbackend.soonganapi.interfaces.report.dto.response.ReportHistoryResponseDto
 import com.soongan.soonganbackend.soonganpersistence.storage.member.MemberEntity
 import com.soongan.soonganbackend.soonganpersistence.storage.report.ReportEntity
 import io.swagger.v3.oas.annotations.media.Schema
@@ -39,29 +40,6 @@ data class MemberInfoResponseDto(
                 profileImageUrl = member.profileImageUrl,
                 selfIntroduction = member.selfIntroduction,
                 reportHistories = reportHistoryResponseDtos
-            )
-        }
-    }
-}
-
-
-@Schema(description = "유저 신고 내역 응답 DTO")
-data class ReportHistoryResponseDto(
-    @Schema(description = "신고 ID", required = true)
-    val id: Long,
-
-    @Schema(description = "신고 대상 ID", required = true)
-    val targetId: Long,
-
-    @Schema(description = "신고 대상 타입 (post/comment)", required = true)
-    val targetType: String,
-) {
-    companion object {
-        fun from(report: ReportEntity): ReportHistoryResponseDto {
-            return ReportHistoryResponseDto(
-                id = report.id!!,
-                targetId = report.targetId,
-                targetType = report.targetType.name,
             )
         }
     }
