+++ apps/soongan-api/src/main/kotlin/com/soongan/soonganbackend/soonganapi/service/report/ReportService.kt
@@ -43,6 +43,8 @@ class ReportService(
             )
         )
 
-        return ReportSaveResponseDto.from(savedReport)
+        val reportHistories = reportAdapter.getReportHistoriesByReportMember(loginMember)
+
+        return ReportSaveResponseDto.from(savedReport, reportHistories)
     }
 }
